Move Kafka notifications of published collections from the Publisher to the PostPublisher, so that Zebedee is aware of the published collection (has moved it into its master foler) when the dp-search-data-extractor requests it (dp-search-data-extractor is triggered by the Kafka message)
